+++ 2003/tjdwlsl1021/app/src/main/java/com/example/myapplication/network/ServerURL.kt
@@ -1,9 +1,6 @@
 package com.example.myapplication.network
 
-/**
- * URL 목록
- * */
-object SearchMovieServerURL {
+object ServerURL {
     const val URL_BASE = "https://openapi.naver.com/"
     const val URL_MOVIE = "v1/search/movie.json"
 }
